Use unittest module for tests
