+++ scripts/run_tests.sh
@@ -1,0 +1,2 @@
+python3 -m pip install .
+python3 -m unittest discover
