+++ test/__init__.py
@@ -1,0 +1,6 @@
+from .test_all import TestStringMethods
+from .test_general import TestGeneral
+from .test_limits import TestLimits
+from .test_pyrtl import TestPyrtl
+from .test_style import TestStyle
+from .test_wires import TestWires
