+++ test/test_all.py
@@ -1,0 +1,11 @@
+from sootty.storage import Wire
+
+import unittest
+
+class TestStringMethods(unittest.TestCase):
+
+    def test_upper(self):
+        self.assertEqual('foo'.upper(), 'FOO')
+
+if __name__ == '__main__':
+    unittest.main()
